check

+++ services/repository/src/test/java/org/sagebionetworks/repo/web/controller/TableControllerAutowireTest.java
@@ -15,8 +15,10 @@ import javax.servlet.ServletException;
 import junit.framework.Assert;
 
 import org.junit.After;
+import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
+import org.sagebionetworks.StackConfiguration;
 import org.sagebionetworks.repo.model.AuthorizationConstants.BOOTSTRAP_PRINCIPAL;
 import org.sagebionetworks.repo.model.Entity;
 import org.sagebionetworks.repo.model.Project;
@@ -47,6 +49,8 @@ public class TableControllerAutowireTest extends AbstractAutowiredControllerTest
 
 	@Autowired
 	private FileHandleDao fileMetadataDao;
+	@Autowired
+	StackConfiguration config;
 
 	private Entity parent;
 	private Long adminUserId;
@@ -55,6 +59,8 @@ public class TableControllerAutowireTest extends AbstractAutowiredControllerTest
 	
 	@Before
 	public void before() throws Exception {
+		Assume.assumeTrue(config.getTableEnabled());
+	
 		adminUserId = BOOTSTRAP_PRINCIPAL.THE_ADMIN_USER.getPrincipalId();
 		
 		parent = new Project();
@@ -65,13 +71,15 @@ public class TableControllerAutowireTest extends AbstractAutowiredControllerTest
 	
 	@After
 	public void after(){
-		if(parent != null){
-			try {
-				servletTestHelper.deleteEntity(dispatchServlet, Project.class, parent.getId(), adminUserId);
-			} catch (Exception e) {} 
-		}
-		for (S3FileHandle handle : handles) {
-			fileMetadataDao.delete(handle.getId());
+		if(config.getTableEnabled()){
+			if(parent != null){
+				try {
+					servletTestHelper.deleteEntity(dispatchServlet, Project.class, parent.getId(), adminUserId);
+				} catch (Exception e) {} 
+			}
+			for (S3FileHandle handle : handles) {
+				fileMetadataDao.delete(handle.getId());
+			}
 		}
 	}
